use strings.HasSuffix
