+++ internal/cloudformation/helpers.go
@@ -1,0 +1,23 @@
+package cloudformation
+
+import (
+	"gopkg.in/yaml.v3"
+)
+
+// findMapNode parses a yaml block and returns the value, when given the key
+func findMapNode(node *yaml.Node, key string) *yaml.Node {
+	if node.Kind == yaml.DocumentNode && len(node.Content) > 0 {
+		node = node.Content[0]
+	}
+	if node.Kind != yaml.MappingNode {
+		return nil
+	}
+	for i := 0; i < len(node.Content); i += 2 {
+		k := node.Content[i]
+		v := node.Content[i+1]
+		if k.Value == key {
+			return v
+		}
+	}
+	return nil
+}
